requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# LangChain & LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1

# Database
sqlalchemy>=2.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
